fix loi kernel

+++ mpp/hal/worker/mpp_device/mpp_device.c
@@ -21,7 +21,6 @@
 #include <errno.h>
 #include <string.h>
 #include <unistd.h>
-#include "linux/types.h"
 
 #include "mpp_env.h"
 #include "mpp_log.h"
@@ -33,11 +32,11 @@
 //https://stackoverflow.com/questions/22496123/what-is-the-meaning-of-this-macro-iormy-macig-0-int
 #define VPU_IOC_MAGIC                       'l'
 
-#define VPU_IOC_SET_CLIENT_TYPE             _IOW(VPU_IOC_MAGIC, 1, __u32)
-#define VPU_IOC_GET_HW_FUSE_STATUS          _IOW(VPU_IOC_MAGIC, 2, __u32)
-#define VPU_IOC_SET_REG                     _IOW(VPU_IOC_MAGIC, 3, __u32)
-#define VPU_IOC_GET_REG                     _IOW(VPU_IOC_MAGIC, 4, __u32)
-#define VPU_IOC_PROBE_IOMMU_STATUS          _IOR(VPU_IOC_MAGIC, 5, __u32)
+#define VPU_IOC_SET_CLIENT_TYPE             _IOW(VPU_IOC_MAGIC, 1, unsigned long)
+#define VPU_IOC_GET_HW_FUSE_STATUS          _IOW(VPU_IOC_MAGIC, 2, unsigned long)
+#define VPU_IOC_SET_REG                     _IOW(VPU_IOC_MAGIC, 3, unsigned long)
+#define VPU_IOC_GET_REG                     _IOW(VPU_IOC_MAGIC, 4, unsigned long)
+#define VPU_IOC_PROBE_IOMMU_STATUS          _IOR(VPU_IOC_MAGIC, 5, unsigned long)
 #define VPU_IOC_WRITE(nr, size)             _IOC(_IOC_WRITE, VPU_IOC_MAGIC, (nr), (size))
 
 typedef struct MppReq_t {
